observation-management-api/app/requirements.txt
Django==3.0.2
psycopg2-binary==2.8.3
djangorestframework
django-shell-plus
django-cors-headers==3.2.1
